refactor(cli): organize runtime modules by domain
