+++ kapsl-runtime/crates/kapsl-cli/src/main.rs
@@ -15,8 +15,6 @@ use kapsl_engine_api::{
     BatchingPolicy, BinaryTensorPacket, Engine, EngineError, EngineHandle, EngineMetrics,
     EngineModelInfo, InferenceRequest, TensorDtype,
 };
-#[cfg(feature = "gpu-device-pool")]
-use kapsl_engine_api::{ExternalDeviceMemory, ExternalDeviceMemoryReport};
 use kapsl_hal::device::DeviceInfo;
 use kapsl_ipc::{IpcServer, TcpServer};
 use kapsl_llm::block_manager::{new_shared_allocator, SharedBlockAllocator};
