+++ kapsl-runtime/crates/kapsl-cli/src/runtime/kv/shared.rs
@@ -1,3 +1,5 @@
+//! In-process KV allocation and live ceiling coordination.
+
 use super::*;
 
 struct KvEngineRecord {
@@ -367,7 +369,7 @@ impl KvCoordinatorInner {
 mod vram_clamp_tests {
     use super::KvCoordinatorInner;
     use crate::app::config::constants::CUDA_DEVICE_MEMORY_LIMIT_ENV;
-    use crate::runtime::device_limits::effective_ceiling_bytes;
+    use crate::runtime::memory::effective_ceiling_bytes;
     use crate::runtime::memory::{
         MemoryAllocationClass, MemoryAuthority, MemoryClaim, MemoryDomain,
     };
