+++ kapsl-runtime/crates/kapsl-cli/src/runtime/memory/device_limits.rs
