+++ kapsl-runtime/crates/kapsl-cli/src/runtime/memory/host.rs
@@ -1,3 +1,5 @@
+//! Host-memory admission and reservation accounting.
+
 use crate::app::config::constants::KAPSL_CPU_MEMORY_LIMIT_MB_ENV;
 use crate::runtime::memory::{MemoryAllocationClass, MemoryClaim, MemoryDomain, MemoryOwner};
 use parking_lot::Mutex;
