+++ kapsl-runtime/crates/kapsl-cli/src/runtime/memory/priority.rs
@@ -1,3 +1,5 @@
+//! Weighted arbitration for memory pressure and admission.
+
 use super::*;
 use std::cmp::Reverse;
 use std::sync::Weak;
