+++ kapsl-runtime/crates/kapsl-cli/src/runtime/serving/mod.rs
@@ -1,0 +1,15 @@
+//! Inference serving, transport, worker, and process supervision.
+
+use super::*;
+
+mod inference;
+mod supervisor;
+mod support;
+mod transport;
+mod worker;
+
+pub(crate) use inference::*;
+pub(crate) use supervisor::*;
+pub(crate) use support::*;
+pub(crate) use transport::*;
+pub(crate) use worker::*;
